Complete test structure
