+++ src/RequestCulture.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Test", "Test\Test.csproj", "{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RequestCulture", "RequestCulture\RequestCulture.csproj", "{0E4F93EB-43D6-4D5A-B91C-8EF984C0297A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebApplication1", "WebApplication1\WebApplication1.csproj", "{8E8FF721-E17A-4D57-B8BB-5C43ACB70306}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0E4F93EB-43D6-4D5A-B91C-8EF984C0297A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0E4F93EB-43D6-4D5A-B91C-8EF984C0297A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0E4F93EB-43D6-4D5A-B91C-8EF984C0297A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0E4F93EB-43D6-4D5A-B91C-8EF984C0297A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8E8FF721-E17A-4D57-B8BB-5C43ACB70306}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8E8FF721-E17A-4D57-B8BB-5C43ACB70306}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8E8FF721-E17A-4D57-B8BB-5C43ACB70306}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8E8FF721-E17A-4D57-B8BB-5C43ACB70306}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E8DFE792-B0F6-4614-96EB-715292457F46}
+	EndGlobalSection
+EndGlobal
