+++ src/Test/WebApplicationFactoryWithHost.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
-namespace test;
+namespace Test;
 
 class WebApplicationFactoryWithHost<TEntryPoint> :
     WebApplicationFactory<TEntryPoint>
