--- test.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "test", "test.csproj", "{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A8D79E4C-0F32-44CB-9DFA-8A5D493EEE5D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
